requirements1.txt
asgiref==3.8.1
astroid==3.3.8
autopep8==2.3.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.7.1
dill==0.3.9
Django==5.1.4
django-environ==0.12.0
django-sendgrid-v5==1.2.4
django_csp==3.8
djangorestframework==3.15.2
djangosaml2==1.9.3
elementpath==4.6.0
idna==3.10
isodate==0.7.2
isort==5.13.2
lxml==5.3.0
MarkupSafe==3.0.2
mccabe==0.7.0
msal==1.31.1
numpy==2.2.0
pandas==2.2.3
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.10.1
pylint==3.3.3
pyOpenSSL==24.3.0
pysaml2==7.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python3-saml==1.16.0
pytz==2024.2
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
tomlkit==0.13.2
tzdata==2024.2
urllib3==2.2.3
xmlschema==2.5.1
xmlsec==1.3.14
